# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Library API
  version: "1.0.0"
  description: Simple library management REST API (BSc project).
servers:
  - url: http://localhost:5000/api
paths:
...
  /reservations:
    post:
      summary: Create reservation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id]
              properties:
                book_id: { type: integer }
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /users/{user_id}/reservations:
    get:
      summary: List reservations for a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [pending, ready, expired, fulfilled, all], default: all }
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /books/{book_id}/reservations:
    get:
      summary: List reservations for a book (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: book_id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /reservations/{reservation_id}/status:
    post:
      summary: Update reservation status (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservation_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, ready, expired, fulfilled]
      responses:
        "200": { description: OK }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reservations/{reservation_id}/cancel:
    post:
      summary: Cancel reservation (user or admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reservation_id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/reservations/expire:
    post:
      summary: Expire overdue reservations (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Expired reservations count
          content:
            application/json:
              schema:
                type: object
                properties:
                  expired_count: { type: integer, minimum: 0 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: invalid_credentials }
        message: { type: string }
        meta:
          type: object
          additionalProperties: true
...